+++ lib/queries/helpers.rb
@@ -1,0 +1,7 @@
+module Queries::Helpers
+
+  def boolean_param(params, attribute)
+    (params[attribute]&.downcase == 'true' ? true : false) if !params[attribute].nil?
+  end
+
+end